Migrated errorHandling as a separate dir. ☘️
