+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
-import { withErrorBoundary } from './ignitus-Shared/Components/errorBoundary';
+import { withErrorBoundary } from './ignitus-Shared/ignitus-ErrorHandlingComponents/errorBoundary';
 import { ProtectedRoutes } from './ignitus-Routes/protectedRoutes';
 // eslint-disable-next-line import/named
 import { PublicRoutes } from './ignitus-Routes/publicRoutes';
