+++ src/ignitus-AboutPage/ignitus-CoreTeam/Components/team.tsx
@@ -3,7 +3,7 @@
 /* eslint-disable jsx-a11y/mouse-events-have-key-events */
 import React from 'react';
 import {Data} from '../constants';
-import {withErrorBoundary} from '../../../ignitus-Shared/Components/errorBoundary';
+import {withErrorBoundary} from '../../../ignitus-Shared/ignitus-ErrorHandlingComponents/errorBoundary';
 import * as S from '../Styles';
 import * as T from '../../../ignitus-Shared/ignitus-DesignSystem/shared';
 import {coreTeam, TeamItemProps} from '../types';
