+++ src/ignitus-Authentication/ignitus-Common-Auth-Student-Prof/Components/commonLoginRegister.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { withErrorBoundary } from '../../../ignitus-Shared/Components/errorBoundary';
+import { withErrorBoundary } from '../../../ignitus-Shared/ignitus-ErrorHandlingComponents/errorBoundary';
 import * as t from './constants';
 import '../Styles/style.scss';
