+++ src/ignitus-Authentication/ignitus-ProfessorSignUp/Components/Signup.jsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import * as t from './Constants';
-import { withErrorBoundary } from '../../../ignitus-Shared/Components/errorBoundary';
+import { withErrorBoundary } from '../../../ignitus-Shared/ignitus-ErrorHandlingComponents/errorBoundary';
 import '../Styles/style.scss';
 
 class Signup extends React.Component {
