+++ src/ignitus-HomePage/ignitus-WhatWeDo/Components/WhatWeProvide.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as S from '../Styles';
 import * as T from '../../../ignitus-Shared/ignitus-DesignSystem/shared';
-import { withErrorBoundary } from '../../../ignitus-Shared/Components/errorBoundary';
+import { withErrorBoundary } from '../../../ignitus-Shared/ignitus-ErrorHandlingComponents/errorBoundary';
 import { data } from '../constants';
 
 const WhatWeProvide = () => (
